+++ app/api/auth/[...all]/route.ts
