+++ app/lib/auth.ts
