+++ app/page.tsx
@@ -1,4 +1,4 @@
-import LoginButton from "@/components/LoginButton";
+import LoginButton from "@/components/LinkedInButton";
 
 export default function Home(){
   return(
